requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
numpy==1.21.5
pandas==1.1.5
python-craigslist==1.1.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.7
